requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyrogram==2.0.106
tgcrypto==1.2.5
python-telegram-bot==21.0
loguru==0.7.2
aiosqlite==0.19.0
python-dateutil==2.8.2
